Refactor key-slot conversion logic
Previously, extract_first_key would perform hashtag extraction on the
key it pulled from command; that meant that the "key" it was returning
might not actually be the key in the command.

This commit refactors things so that

* extract_first_key actually extracts the first key
* KeySlotConverter.convert does hashtag extraction
* Router's find_node_key and find_primary_node_key can now be
  implemented in terms of a function "find_node_by_key", because they
  _actually_ get the key from the command.
